mnaully merged in PR #52

+++ lib/src/money.dart
@@ -100,8 +100,8 @@ class Money implements Comparable<Money> {
   ///
   factory Money.from(num amount, Currency currency) {
     final minorUnits = BigInt.from(
-        (amount * currency.precisionFactor.toInt() + (amount >= 0 ? 0.5 : -0.5))
-            .toInt());
+      amount * currency.precisionFactor.toInt() + (amount >= 0 ? 0.5 : -0.5),
+    );
 
     return Money._from(MinorUnits.from(minorUnits), currency);
   }
